Money entry fields render committed values as $-grouped (owner gate catch)
Gate-testing the C1 fix caught the seam it left: the Salary field showed
bare '2000' directly between the grouped '$1,500' Balance and the
'Collect $2,000 Salary' button. The one shared formatter (renamed
NumberFormatter.money, file IntegerFormatter.swift -> MoneyFormatter.swift)
now renders committed values grouped with a $ prefix while typing stays
plain digits, and parses leniently ('$2,0005' -> 20005; empty still nil,
preserving clear-to-unset) — empirically verified round-trips. All eight
money TextFields retargeted; AddNewPlayerView's two Text("$") prefix
shims removed (the formatter carries the $ now); AGENTS.md listing
updated. Owner-approved design: $-grouped on commit, one formatter (#37's
single-formatter rule intact).

+++ iBanker/View/GameModeSection.swift
@@ -108,7 +108,7 @@ struct GameModeSection: View {
                 HStack {
                     Text("Default Balance")
                     Spacer()
-                    TextField("Initial Balance", value: customBalanceBinding, formatter: NumberFormatter.integer)
+                    TextField("Initial Balance", value: customBalanceBinding, formatter: NumberFormatter.money)
                         .keyboardType(.numberPad)
                         .autocorrectionDisabled()
                         .multilineTextAlignment(.trailing)
@@ -117,7 +117,7 @@ struct GameModeSection: View {
                 HStack {
                     Text("Default Salary")
                     Spacer()
-                    TextField("Initial Salary", value: customSalaryBinding, formatter: NumberFormatter.integer)
+                    TextField("Initial Salary", value: customSalaryBinding, formatter: NumberFormatter.money)
                         .keyboardType(.numberPad)
                         .autocorrectionDisabled()
                         .multilineTextAlignment(.trailing)
